WW-5220 Moves JUnit related test into the JUnit plugin Also moves all the plugin classes under junit package
