+++ plugins/jfreechart/src/test/java/org/apache/struts2/dispatcher/ChartResultTest.java
@@ -23,7 +23,7 @@ import com.opensymphony.xwork2.ActionInvocation;
 import com.opensymphony.xwork2.ActionProxy;
 import com.opensymphony.xwork2.util.ValueStack;
 import org.apache.struts2.ServletActionContext;
-import org.apache.struts2.StrutsTestCase;
+import org.apache.struts2.junit.StrutsTestCase;
 import org.easymock.EasyMock;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.JFreeChart;
@@ -36,6 +36,7 @@ import java.io.IOException;
 
 
 /**
+ *
  */
 public class ChartResultTest extends StrutsTestCase {
 
@@ -50,7 +51,7 @@ public class ChartResultTest extends StrutsTestCase {
     public void testChart() throws Exception {
         EasyMock.expect(responseMock.getOutputStream()).andReturn(os);
         EasyMock.replay(responseMock, mockActionProxy, actionInvocation);
-        
+
         ChartResult result = new ChartResult();
 
         result.setChart(mockChart);
@@ -62,7 +63,7 @@ public class ChartResultTest extends StrutsTestCase {
         EasyMock.verify(responseMock);
         assertTrue(os.isWritten());
     }
-    
+
     public void testContentTypePng() throws Exception {
         EasyMock.expect(responseMock.getOutputStream()).andReturn(os);
         responseMock.setContentType("image/png");
@@ -79,7 +80,7 @@ public class ChartResultTest extends StrutsTestCase {
         EasyMock.verify(responseMock);
         assertTrue(os.isWritten());
     }
-    
+
     public void testContentTypeJpg() throws Exception {
         EasyMock.expect(responseMock.getOutputStream()).andReturn(os);
         responseMock.setContentType("image/jpg");
@@ -101,7 +102,7 @@ public class ChartResultTest extends StrutsTestCase {
     public void testChartNotSet() {
         ChartResult result = new ChartResult();
         EasyMock.replay(responseMock, mockActionProxy, actionInvocation);
-        
+
         // expect exception if chart not set.
         result.setChart(null);
 
@@ -141,7 +142,7 @@ public class ChartResultTest extends StrutsTestCase {
         assertEquals("150", result.getWidth().toString());
         assertTrue(os.isWritten());
     }
-    
+
     protected void setUp() throws Exception {
         super.setUp();
 
@@ -160,8 +161,8 @@ public class ChartResultTest extends StrutsTestCase {
         actionInvocation = EasyMock.createMock(ActionInvocation.class);
 
         EasyMock.expect(actionInvocation.getStack()).andReturn(stack).anyTimes();
-        
-        
+
+
         os = new MockServletOutputStream();
         responseMock = EasyMock.createNiceMock(HttpServletResponse.class);
