+++ plugins/junit/src/test/java/org/apache/struts2/junit/MySessionBean.java
@@ -16,18 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2.session;
+package org.apache.struts2.junit;
 
-import com.opensymphony.xwork2.ActionSupport;
+public class MySessionBean {
 
-/**
- * An action which reads a value from the session.
- * The value is previously set by SessionSetAction.
- */
-public class SessionGetAction extends ActionSupport {
-    private static final long serialVersionUID = 8366502863472148631L;
-
-    public String execute() {
-        return ActionSupport.SUCCESS;
+    public void doSomething() {
+        // do Something
     }
 }
