+++ plugins/junit/src/test/java/org/apache/struts2/junit/StrutsJUnit4TestCaseTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2;
+package org.apache.struts2.junit;
 
 import com.opensymphony.xwork2.ActionProxy;
 import org.junit.Assert;
@@ -27,7 +27,7 @@ import org.junit.Test;
  * Date: 8/15/11
  * Time: 7:04 PM
  */
-public class StrutsJUnit4TestCaseTest extends StrutsJUnit4TestCase<JUnitTestAction>{
+public class StrutsJUnit4TestCaseTest extends StrutsJUnit4TestCase<JUnitTestAction> {
     @Test
     public void testExecuteActionAgainstCustomStrutsConfigFile() throws Exception {
         String output = executeAction("/test/testAction-2.action");
@@ -38,7 +38,7 @@ public class StrutsJUnit4TestCaseTest extends StrutsJUnit4TestCase<JUnitTestActi
     public void testSessionInitialized() throws Exception {
         ActionProxy proxy = getActionProxy("/test/testAction-2.action");
         Assert.assertNotNull("invocation session should being initialized",
-                proxy.getInvocation().getInvocationContext().getSession());
+            proxy.getInvocation().getInvocationContext().getSession());
     }
 
     @Override
