+++ plugins/junit/src/test/resources/struts.xml
@@ -20,13 +20,13 @@
  */
 -->
 <!DOCTYPE struts PUBLIC
-    "-//Apache Software Foundation//DTD Struts Configuration 2.1.7//EN"
-    "http://struts.apache.org/dtds/struts-2.1.dtd">
+        "-//Apache Software Foundation//DTD Struts Configuration 2.1.7//EN"
+        "http://struts.apache.org/dtds/struts-2.1.dtd">
 
 <struts>
     <constant name="struts.objectFactory" value="spring"/>
     <package name="test" namespace="/test" extends="struts-default">
-        <action name="testAction" class="org.apache.struts2.JUnitTestAction">
+        <action name="testAction" class="org.apache.struts2.junit.JUnitTestAction">
             <result type="freemarker">/template-1.ftl</result>
         </action>
     </package>
